clippy: workspace auto-fix sweep + bump to 0.1.3
Apply `cargo clippy --fix --workspace --all-targets` to clear the
mechanical clippy debt across the workspace. 27 source files touched
across 11 crates. The auto-fix handles every lint clippy classifies
as machine-applicable, which covers the high-frequency lints
documented in the cleanup plan (collapsible_if, unwrap_or_default,
unnecessary_sort_by) and opportunistically clears a few of the
long-tail lints as well. Tests pass workspace-wide.

Also fixes `crates/astraea-graph/benches/traversal_bench.rs`, which
contained an empty `criterion_main!()` invocation that fails to
compile under `cargo clippy --all-targets`. Replaced with a no-op
`fn main()` until real bench groups are added. Without this, the
`Clippy astraea-graph` job in the version-gate can never reach the
lint stage on any PR.

Adds `docs/clippy-cleanup-plan.md` to `.gitignore` — the cleanup
plan is a working tracking doc kept locally, not part of the
distributed docs.

Per-crate clippy debt remains in astraea-gnn, astraea-gpu,
astraea-query, astraea-rag, and astraea-storage. These require
manual fixes (approx_constant false positives, type_complexity in
trait bounds, residual needless_range_loop, etc.) and land as
Wave 2 per the cleanup plan.

+++ crates/astraea-cluster/src/partition.rs
@@ -38,7 +38,7 @@ fn fnv1a_u64(mut input: u64) -> u64 {
     const FNV_PRIME: u64 = 0x100000001b3;
     let mut hash: u64 = FNV_OFFSET;
     for _ in 0..8 {
-        let byte = (input & 0xff) as u64;
+        let byte = input & 0xff;
         hash ^= byte;
         hash = hash.wrapping_mul(FNV_PRIME);
         input >>= 8;
